refactor(chat): split agent-chat-view.tsx into utils + hooks
Pure structural decomposition, zero behavior change. agent-chat-view.tsx
3324 -> 2797 lines. Extracted:
- chat-message-utils.ts: pure fns (sort/merge/buildTimeline/computeGroupPositions/
  group helpers/pointer fns/useLatest)
- chat-view-parts.tsx: leaf components (NapSeparator/ArtifactCard/MentionHighlight)
- hooks: useMessageFlags, useChatSheets, useFileAttachments, useTextSelectionQuote

Moves are verbatim: dep arrays, effect order, ref ownership, and render tree
unchanged. The orchestration core (load/WS/polling/send) and the 3 hooks that
co-own its shared refs (composer-draft, scroll, nap/mark-read) intentionally
left in the component. Verified behavior-preserved by code review x2 + QA browser
pass (cache-first/no-flash, no-stuck-skeleton, send-dm AC4, presence all green).
