+++ src/web/src/components/agent-chat/agent-chat-view.tsx
@@ -27,12 +27,8 @@ import {
   getActiveTask,
   retryTask,
   markInboxRead,
-  getIssue,
-  getTrace,
   updateIssue,
   listFlaggedMessageIds,
-  flagMessage as apiFlagMessage,
-  unflagMessage as apiUnflagMessage,
   getAgentSkills,
 } from "@/lib/api";
 import {
@@ -52,12 +48,21 @@ import {
   shouldSkipFastLoad,
   markFastLoadCompleted,
 } from "@/components/agent-chat/fast-load-gate";
-import type { PreviousConversation, TraceTask } from "@/lib/api";
+import {
+  sortMessages,
+  mergeMessages,
+  computeGroupPositions,
+  buildTimeline,
+  shouldPersistPointerForLoad,
+  pointerRefreshTargetForTaskCreated,
+  useLatest,
+} from "@/components/agent-chat/chat-message-utils";
+import type { NapMarker } from "@/components/agent-chat/chat-message-utils";
+import type { PreviousConversation } from "@/lib/api";
 import type {
   Artifact,
   Conversation,
   Issue,
-  IssueComment,
   Message,
   SkillEntry,
   TaskApi as Task,
@@ -66,7 +71,6 @@ import type {
 } from "@alook/shared";
 import { useAgentContext } from "@/contexts/agent-context";
 import { useInboxCount } from "@/contexts/inbox-count-context";
-import { useFlagCount } from "@/contexts/flag-count-context";
 import { useChannel } from "@/contexts/channel-context";
 import { toast } from "sonner";
 import { cn } from "@/lib/utils";
@@ -88,6 +92,10 @@ import {
   X,
 } from "lucide-react";
 import { useCachedMessages } from "@/hooks/use-cached-messages";
+import { useMessageFlags } from "@/hooks/use-message-flags";
+import { useChatSheets } from "@/hooks/use-chat-sheets";
+import { useFileAttachments } from "@/hooks/use-file-attachments";
+import { useTextSelectionQuote } from "@/hooks/use-text-selection-quote";
 import { useSlashCommand } from "@/hooks/use-slash-command";
 import { SlashCommandPopup } from "@/components/agent-chat/slash-command-popup";
 import {
@@ -109,10 +117,14 @@ import {
   computeArtifactVersions,
 } from "@/components/artifact-content-renderer";
 import { ScrollToBottomButton } from "@/components/ui/scroll-to-bottom-button";
-import { MessageItem, SystemCard, AgentRow } from "@/components/agent-chat/message-list";
+import { MessageItem, AgentRow } from "@/components/agent-chat/message-list";
 import { PresenceLine } from "@/components/agent-chat/presence-line";
 import { parseAvatarUrl } from "@/components/avatar";
-import { AgentPreviewCard } from "@/components/agent-preview-card";
+import {
+  MENTION_COMPONENTS,
+  NapSeparator,
+  ArtifactCard,
+} from "@/components/agent-chat/chat-view-parts";
 import {
   Popover,
   PopoverTrigger,
@@ -121,375 +133,6 @@ import {
 
 const MESSAGE_LIMIT = 20;
 const MAX_CONV_FETCHES_PER_CLICK = 5;
-const MAX_FILE_SIZE = 10 * 1024 * 1024; // 10 MB
-
-function MentionHighlight(
-  props: Record<string, unknown> & { children?: React.ReactNode },
-) {
-  const { children, ...rest } = props;
-  const { agents } = useAgentContext();
-  const agentId = (rest["data-agent-id"] ?? rest.dataAgentId) as
-    | string
-    | undefined;
-  let agent = agentId ? agents.find((a) => a.id === agentId) : undefined;
-  if (!agent && typeof children === "string") {
-    const nameToMatch = children.startsWith("@") ? children.slice(1) : children;
-    agent = agents.find(
-      (a) => a.name.toLowerCase() === nameToMatch.toLowerCase(),
-    );
-  }
-  if (agent) {
-    return (
-      <Popover>
-        <PopoverTrigger
-          openOnHover
-          delay={300}
-          nativeButton={false}
-          render={<span className="mention-highlight cursor-pointer" />}
-        >
-          {children}
-        </PopoverTrigger>
-        <PopoverContent side="top" className="w-fit max-w-80">
-          <AgentPreviewCard agent={agent} />
-        </PopoverContent>
-      </Popover>
-    );
-  }
-  return <span className="mention-highlight">{children}</span>;
-}
-const MENTION_COMPONENTS: Record<
-  string,
-  React.ComponentType<Record<string, unknown> & { children?: React.ReactNode }>
-> = {
-  mention: MentionHighlight,
-  p: ({
-    children,
-    node,
-    ...rest
-  }: Record<string, unknown> & { children?: React.ReactNode }) => {
-    void node;
-    return (
-      <div data-md-p="" {...rest}>
-        {children}
-      </div>
-    );
-  },
-};
-
-type EventIconType = "issue" | "email" | "calendar";
-
-export function getEventIconType(
-  content: string,
-  conversationType?: string | null,
-): EventIconType {
-  if (conversationType === "issue_event") return "issue";
-  if (conversationType === "email_notification") return "email";
-  if (conversationType === "calendar_event") return "calendar";
-
-  const lower = content.toLowerCase();
-  if (lower.startsWith("issue ") || lower.startsWith("issue:")) return "issue";
-  if (lower.includes("email")) return "email";
-  return "calendar";
-}
-
-/**
- * Classify a system-event message by its metadata resource id first (reliable,
- * stamped at all creation sites), falling back to the content/conversationType
- * heuristic only when no id is present. Drives both the card icon and label.
- */
-export function eventTypeFromMessage(
-  metadata: Record<string, unknown> | null | undefined,
-  content: string,
-  conversationType?: string | null,
-): EventIconType {
-  if (metadata?.issueId) return "issue";
-  if (metadata?.emailId) return "email";
-  if (metadata?.calendarEventId) return "calendar";
-  return getEventIconType(content, conversationType);
-}
-
-/** Sort messages by (created_at, id) ascending — guarantees chronological order. */
-export function sortMessages(msgs: Message[]): Message[] {
-  return msgs.slice().sort((a, b) => {
-    const cmp = a.created_at.localeCompare(b.created_at);
-    if (cmp !== 0) return cmp;
-    return a.id.localeCompare(b.id);
-  });
-}
-
-/** Merge two message arrays by ID (latest wins), then sort chronologically. */
-export function mergeMessages(
-  existing: Message[],
-  incoming: Message[],
-): Message[] {
-  const merged = new Map<string, Message>();
-  for (const m of existing) merged.set(m.id, m);
-  for (const m of incoming) merged.set(m.id, m);
-  return sortMessages([...merged.values()]);
-}
-
-export type NapMarker = { agentName: string; created_at: string; id: string };
-
-type TimelineItem =
-  | { kind: "message"; data: Message }
-  | { kind: "artifact"; data: Artifact }
-  | { kind: "nap"; data: NapMarker };
-
-export type GroupPosition = "first" | "middle" | "last" | "solo";
-
-// Which conversational SIDE a timeline item belongs to. Items on the same side
-// that are adjacent (within the time threshold, no nap between) form one
-// Slack/Discord-style cluster sharing a single avatar + name header. The agent
-// side intentionally includes event cards (email/issue/calendar) and artifacts
-// (files), so a run like "card → reply → card → reply" reads as ONE Gwennie
-// cluster with one header — not four separate avatars.
-type GroupSide = "user" | "agent" | null;
-
-function groupSideOf(item: TimelineItem): GroupSide {
-  if (item.kind === "artifact") return "agent"; // files come from the agent
-  if (item.kind !== "message") return null; // nap markers break clusters
-  const role = item.data.role;
-  if (role === "user") return "user";
-  if (role === "assistant" || role === "event") return "agent";
-  return null;
-}
-
-function itemCreatedAt(item: TimelineItem): string | undefined {
-  return item.data.created_at;
-}
-
-export function computeGroupPositions(
-  timeline: TimelineItem[],
-): (GroupPosition | null)[] {
-  const positions: (GroupPosition | null)[] = new Array(timeline.length).fill(
-    null,
-  );
-  const GROUP_THRESHOLD_MS = 60_000;
-
-  const adjacentSameCluster = (a: TimelineItem | null, b: TimelineItem | null) => {
-    if (!a || !b) return false;
-    const sa = groupSideOf(a);
-    const sb = groupSideOf(b);
-    if (sa === null || sa !== sb) return false;
-    const ta = itemCreatedAt(a);
-    const tb = itemCreatedAt(b);
-    if (!ta || !tb) return true; // missing timestamp → don't split the cluster
-    return Math.abs(new Date(ta).getTime() - new Date(tb).getTime()) < GROUP_THRESHOLD_MS;
-  };
-
-  for (let i = 0; i < timeline.length; i++) {
-    const item = timeline[i];
-    if (groupSideOf(item) === null) continue; // nap → leave null
-
-    const prev = i > 0 ? timeline[i - 1] : null;
-    const next = i < timeline.length - 1 ? timeline[i + 1] : null;
-
-    const sameAsPrev = adjacentSameCluster(prev, item);
-    const sameAsNext = adjacentSameCluster(item, next);
-
-    if (sameAsPrev && sameAsNext) positions[i] = "middle";
-    else if (sameAsPrev && !sameAsNext) positions[i] = "last";
-    else if (!sameAsPrev && sameAsNext) positions[i] = "first";
-    else positions[i] = "solo";
-  }
-
-  return positions;
-}
-
-export function buildTimeline(
-  messages: Message[],
-  artifacts: Artifact[],
-  napMarkers: NapMarker[],
-  currentConversationId?: string | null,
-): TimelineItem[] {
-  if (!currentConversationId || napMarkers.length === 0) {
-    const items: TimelineItem[] = [
-      ...messages.map((m): TimelineItem => ({ kind: "message", data: m })),
-      ...artifacts.map((a): TimelineItem => ({ kind: "artifact", data: a })),
-      ...napMarkers.map((n): TimelineItem => ({ kind: "nap", data: n })),
-    ];
-    items.sort((a, b) => {
-      const cmp = a.data.created_at.localeCompare(b.data.created_at);
-      if (cmp !== 0) return cmp;
-      if (a.kind === "nap" || b.kind === "nap") {
-        if (a.kind === "nap" && b.kind !== "nap") return 1;
-        if (a.kind !== "nap" && b.kind === "nap") return -1;
-      }
-      if (a.kind !== b.kind) return a.kind === "message" ? -1 : 1;
-      return a.data.id.localeCompare(b.data.id);
-    });
-    // Strictly chronological — a file uploaded mid-task stays where it appeared
-    // and the agent's reply follows below it in the same cluster (no reorder, so
-    // nothing jumps when the reply lands). Gus's call, 2026-06-01.
-    return items;
-  }
-
-  const napConvIds = new Set(napMarkers.map((n) => n.id.replace(/^nap-/, "")));
-  const sortedNaps = [...napMarkers].sort((a, b) =>
-    a.created_at.localeCompare(b.created_at),
-  );
-
-  const groupItems = (convId: string): TimelineItem[] => {
-    const msgs: TimelineItem[] = messages
-      .filter((m) => m.conversation_id === convId)
-      .map((m) => ({ kind: "message" as const, data: m }));
-    const arts: TimelineItem[] = artifacts
-      .filter((a) => a.conversation_id === convId)
-      .map((a) => ({ kind: "artifact" as const, data: a }));
-    const sorted = [...msgs, ...arts].sort((a, b) => {
-      const cmp = a.data.created_at.localeCompare(b.data.created_at);
-      if (cmp !== 0) return cmp;
-      if (a.kind !== b.kind) return a.kind === "message" ? -1 : 1;
-      return a.data.id.localeCompare(b.data.id);
-    });
-    return sorted; // chronological, no artifact reorder (see above)
-  };
-
-  const result: TimelineItem[] = [];
-
-  for (const nap of sortedNaps) {
-    const convId = nap.id.replace(/^nap-/, "");
-    result.push(...groupItems(convId));
-    result.push({ kind: "nap", data: nap });
-  }
-
-  result.push(...groupItems(currentConversationId));
-
-  const knownConvIds = new Set([...napConvIds, currentConversationId]);
-  const orphanMsgs = messages.filter(
-    (m) => !knownConvIds.has(m.conversation_id),
-  );
-  const orphanArts = artifacts.filter(
-    (a) => !knownConvIds.has(a.conversation_id),
-  );
-  if (orphanMsgs.length > 0 || orphanArts.length > 0) {
-    const orphanItems: TimelineItem[] = [
-      ...orphanMsgs.map((m): TimelineItem => ({ kind: "message", data: m })),
-      ...orphanArts.map((a): TimelineItem => ({ kind: "artifact", data: a })),
-    ];
-    orphanItems.sort((a, b) => {
-      const cmp = a.data.created_at.localeCompare(b.data.created_at);
-      if (cmp !== 0) return cmp;
-      if (a.kind !== b.kind) return a.kind === "message" ? -1 : 1;
-      return a.data.id.localeCompare(b.data.id);
-    });
-    const napIdx = result.findIndex((item) => item.kind === "nap");
-    if (napIdx >= 0) {
-      result.splice(napIdx, 0, ...orphanItems);
-    } else {
-      result.push(...orphanItems);
-    }
-  }
-
-  return result;
-}
-
-/**
- * Whether a completed conversation load may persist the per-channel `last_open`
- * pointer. Only the SLOW path (param-less open, `targetConvId` absent) resolves
- * the channel's latest-created conversation; the FAST path (`?conv=<id>`) loads
- * an explicit, possibly old conversation and must NOT touch the pointer (doing
- * so reintroduces the wrong-conversation flash). See {@link LastOpenEntry}.
- *
- * Pure mirror of the inline Phase B write guard, exported for unit testing the
- * fast-vs-slow decision without rendering the component.
- */
-export function shouldPersistPointerForLoad(
-  targetConvId: string | null | undefined,
-): boolean {
-  return !targetConvId;
-}
-
-/**
- * Decide whether a `task.created` WS event should refresh this view's
- * per-channel `last_open` pointer, and to which conversation.
- *
- * The `last_open` pointer carries "latest-created conversation for this
- * agent+channel" semantics (see {@link setLastOpenConversation}). A
- * `task.created` event is the client learning of a (possibly newer)
- * conversation in real time, so it is a valid moment to refresh the pointer —
- * but ONLY when the event belongs to exactly this agent + active channel.
- *
- * Returns the conversation id to point at, or `null` to skip the write
- * (different agent, different channel, or the pointer already points there).
- *
- * Pure so the scoping logic is unit-testable in the node-env web test suite
- * (no component render). The caller supplies `serverMessageCount` separately
- * (A1: derived from the locally-cached message count) when it performs the
- * actual `setLastOpenConversation` write.
- */
-export function pointerRefreshTargetForTaskCreated(args: {
-  /** The TaskApi from the `task.created` event. */
-  task: Pick<Task, "agent_id" | "channel" | "conversation_id">;
-  /** The agent this view is rendering. */
-  agentId: string;
-  /** The active channel string (never null — null maps to "default"). */
-  activeChannel: string;
-  /** The conversation id the pointer currently references, if known. */
-  currentPointerConvId: string | null;
-}): string | null {
-  const { task, agentId, activeChannel, currentPointerConvId } = args;
-  // Scope guard: never touch another agent's or channel's pointer. The event's
-  // null channel normalizes to "default" to match the UI's channel string
-  // (channel-context.tsx). activeChannel is always a non-null string.
-  if (task.agent_id !== agentId) return null;
-  if ((task.channel ?? "default") !== activeChannel) return null;
-  // Already pointing here — nothing to do.
-  if (task.conversation_id === currentPointerConvId) return null;
-  return task.conversation_id;
-}
-
-function useLatest<T>(value: T) {
-  const ref = useRef(value);
-  useEffect(() => {
-    ref.current = value;
-  }, [value]);
-  return ref;
-}
-
-function NapSeparator({ agentName }: { agentName: string }) {
-  return (
-    <div className="flex items-center gap-3 py-4 select-none" aria-hidden>
-      <div className="flex-1 border-t border-border/40" />
-      <span className="text-xs text-muted-foreground/60 whitespace-nowrap">
-        {agentName} took a nap 💤
-      </span>
-      <div className="flex-1 border-t border-border/40" />
-    </div>
-  );
-}
-
-// Agent-side artifact (file) card — same shared SystemCard shell + card
-// language as the event cards (locked prototype: icon chip / FILE / filename /
-// size). The version badge rides along as the title's trailing element.
-function ArtifactCard({
-  artifact,
-  version,
-  hasDuplicates,
-  onClick,
-}: {
-  artifact: Artifact;
-  version: number;
-  hasDuplicates: boolean;
-  onClick: (a: Artifact) => void;
-}) {
-  return (
-    <SystemCard
-      icon={FileText}
-      label="FILE"
-      title={artifact.filename}
-      preview={formatSize(artifact.size)}
-      trailing={
-        hasDuplicates ? (
-          <span className="shrink-0 text-xs text-muted-foreground bg-muted rounded-full px-1.5 py-0.5 font-normal">
-            v{version}
-          </span>
-        ) : undefined
-      }
-      onClick={() => onClick(artifact)}
-    />
-  );
-}
 
 export function AgentChatView({
   agentId: propAgentId,
@@ -559,30 +202,46 @@ export function AgentChatView({
   const [hasMore, setHasMore] = useState(false);
   const [loadingMore, setLoadingMore] = useState(false);
   const [artifacts, setArtifacts] = useState<Artifact[]>([]);
-  const [artifactSheetOpen, setArtifactSheetOpen] = useState(false);
-  const [selectedArtifact, setSelectedArtifact] = useState<Artifact | null>(
-    null,
-  );
-  const [emailSheetOpen, setEmailSheetOpen] = useState(false);
-  const [selectedEmailId, setSelectedEmailId] = useState<string | null>(null);
-  const [calendarEventSheetOpen, setCalendarEventSheetOpen] = useState(false);
-  const [selectedCalendarEventId, setSelectedCalendarEventId] = useState<
-    string | null
-  >(null);
-  const [issueSheetOpen, setIssueSheetOpen] = useState(false);
-  const [selectedIssueId, setSelectedIssueId] = useState<string | null>(null);
-  const [issueDetail, setIssueDetail] = useState<{
-    issue: Issue & { trace_id?: string | null };
-    messages: Message[];
-    comments: IssueComment[];
-    artifacts: Artifact[];
-  } | null>(null);
-  const [issueDetailLoading, setIssueDetailLoading] = useState(false);
-  const [issueTraceTasks, setIssueTraceTasks] = useState<TraceTask[] | null>(
-    null,
-  );
-  const [issueActiveTask, setIssueActiveTask] = useState<Task | null>(null);
-  const [pendingFiles, setPendingFiles] = useState<File[]>([]);
+  const {
+    artifactSheetOpen,
+    setArtifactSheetOpen,
+    selectedArtifact,
+    setSelectedArtifact,
+    emailSheetOpen,
+    setEmailSheetOpen,
+    selectedEmailId,
+    setSelectedEmailId,
+    calendarEventSheetOpen,
+    setCalendarEventSheetOpen,
+    selectedCalendarEventId,
+    setSelectedCalendarEventId,
+    issueSheetOpen,
+    setIssueSheetOpen,
+    selectedIssueId,
+    setSelectedIssueId,
+    issueDetail,
+    setIssueDetail,
+    issueDetailLoading,
+    issueTraceTasks,
+    issueActiveTask,
+    setIssueActiveTask,
+    openIssue,
+    issueConvId,
+    issueTaskId,
+  } = useChatSheets(workspaceId);
+  const {
+    pendingFiles,
+    setPendingFiles,
+    fileInputRef,
+    addPendingFiles,
+    handleFileSelect,
+    removePendingFile,
+    dragging,
+    handleDragEnter,
+    handleDragLeave,
+    handleDragOver,
+    handleDrop,
+  } = useFileAttachments();
   const [caretIndex, setCaretIndex] = useState<number | null>(null);
   const [previousConversations, setPreviousConversations] = useState<
     PreviousConversation[]
@@ -614,22 +273,8 @@ export function AgentChatView({
     }
   });
   const [isMultiLine, setIsMultiLine] = useState(false);
-  const [selectionPopup, setSelectionPopup] = useState<{
-    text: string;
-    x: number;
-    y: number;
-  } | null>(null);
-  const [flaggedIds, setFlaggedIds] = useState<Set<string>>(new Set());
-  const flaggedIdsRef = useRef(flaggedIds);
-  useEffect(() => {
-    flaggedIdsRef.current = flaggedIds;
-  });
-
-  const {
-    increment: flagIncrement,
-    decrement: flagDecrement,
-    refresh: flagRefresh,
-  } = useFlagCount();
+  const { flaggedIds, setFlaggedIds, handleToggleFlag } =
+    useMessageFlags(workspaceId);
 
   const { writeToCache } = useCachedMessages(targetConvId ?? null, workspaceId);
   const writeToCacheRef = useRef(writeToCache);
@@ -725,7 +370,6 @@ export function AgentChatView({
   // skeleton if a run is cancelled mid-flight. See fast-load-gate.ts (Part 2-a /
   // TODO 6 + stuck-skeleton fix).
   const fastLoadGateRef = useRef(createFastLoadGateState());
-  const fileInputRef = useRef<HTMLInputElement>(null);
   // TipTap composer imperative handle (focus / clear / isEmpty / anchor coords).
   const composerRef = useRef<ChatComposerHandle>(null);
 
@@ -2163,68 +1807,7 @@ export function AgentChatView({
     });
   }, [subscribeReconnect, conversation?.id, workspaceId]);
 
-  // Validate (10 MB cap) + stage files. Shared by the file picker, drag/drop,
-  // and the composer's paste/drop handlers.
-  const addPendingFiles = useCallback((files: File[]) => {
-    const valid: File[] = [];
-    for (const file of files) {
-      if (file.size > MAX_FILE_SIZE) {
-        toast.error(`"${file.name}" exceeds 10 MB limit`);
-      } else {
-        valid.push(file);
-      }
-    }
-    if (valid.length > 0) {
-      setPendingFiles((prev) => [...prev, ...valid]);
-    }
-  }, []);
-
-  const handleFileSelect = useCallback(
-    (e: React.ChangeEvent<HTMLInputElement>) => {
-      const fileList = e.target.files;
-      if (!fileList) return;
-      addPendingFiles(Array.from(fileList));
-      // Reset input so re-selecting the same file works
-      e.target.value = "";
-    },
-    [addPendingFiles],
-  );
-
-  const removePendingFile = useCallback((index: number) => {
-    setPendingFiles((prev) => prev.filter((_, i) => i !== index));
-  }, []);
-
-  const handleTextSelect = useCallback(() => {
-    const selection = window.getSelection();
-    if (!selection || selection.isCollapsed || !selection.rangeCount) {
-      setSelectionPopup(null);
-      return;
-    }
-    const text = selection.toString().trim();
-    if (!text) {
-      setSelectionPopup(null);
-      return;
-    }
-    // Only allow quoting from assistant message bubbles
-    const range = selection.getRangeAt(0);
-    const container =
-      range.commonAncestorContainer instanceof HTMLElement
-        ? range.commonAncestorContainer
-        : range.commonAncestorContainer.parentElement;
-    if (!container?.closest("[data-quote-source]")) {
-      setSelectionPopup(null);
-      return;
-    }
-    const rects = range.getClientRects();
-    const lastRect = rects[rects.length - 1] || range.getBoundingClientRect();
-    setSelectionPopup({ text, x: lastRect.right, y: lastRect.top - 4 });
-  }, []);
-
-  useEffect(() => {
-    document.addEventListener("selectionchange", handleTextSelect);
-    return () =>
-      document.removeEventListener("selectionchange", handleTextSelect);
-  }, [handleTextSelect]);
+  const { selectionPopup, setSelectionPopup } = useTextSelectionQuote();
 
   const handleQuoteSelection = useCallback(() => {
     if (selectionPopup) {
@@ -2234,43 +1817,6 @@ export function AgentChatView({
       composerRef.current?.focus();
     }
   }, [selectionPopup]);
-
-  const [dragging, setDragging] = useState(false);
-  const dragCounter = useRef(0);
-
-  const handleDragEnter = useCallback((e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    dragCounter.current++;
-    if (e.dataTransfer.types.includes("Files")) {
-      setDragging(true);
-    }
-  }, []);
-
-  const handleDragLeave = useCallback((e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    dragCounter.current--;
-    if (dragCounter.current === 0) {
-      setDragging(false);
-    }
-  }, []);
-
-  const handleDragOver = useCallback((e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-  }, []);
-
-  const handleDrop = useCallback(
-    (e: React.DragEvent) => {
-      e.preventDefault();
-      e.stopPropagation();
-      setDragging(false);
-      dragCounter.current = 0;
-      addPendingFiles(Array.from(e.dataTransfer.files));
-    },
-    [addPendingFiles],
-  );
 
   const handleSend = async () => {
     const rawContent = input.trim();
@@ -2468,79 +2014,6 @@ export function AgentChatView({
     setTaskMessages([]);
     startPolling(newTask.id, conversation.id);
   }, [activeTask, conversation, workspaceId, startPolling]);
-
-  const handleToggleFlag = useCallback(
-    async (messageId: string) => {
-      const wasFlagged = flaggedIdsRef.current.has(messageId);
-      setFlaggedIds((prev) => {
-        const next = new Set(prev);
-        if (wasFlagged) next.delete(messageId);
-        else next.add(messageId);
-        return next;
-      });
-      if (wasFlagged) {
-        flagDecrement();
-        apiUnflagMessage(workspaceId, messageId)
-          .then(() => {
-            flagRefresh();
-          })
-          .catch(() => {
-            setFlaggedIds((prev) => new Set(prev).add(messageId));
-            flagIncrement();
-          });
-      } else {
-        flagIncrement();
-        apiFlagMessage(workspaceId, messageId)
-          .then(() => {
-            flagRefresh();
-          })
-          .catch(() => {
-            setFlaggedIds((prev) => {
-              const next = new Set(prev);
-              next.delete(messageId);
-              return next;
-            });
-            flagDecrement();
-          });
-      }
-    },
-    [workspaceId, flagIncrement, flagDecrement, flagRefresh],
-  );
-
-  const openIssue = useCallback(
-    async (issueId: string) => {
-      setSelectedIssueId(issueId);
-      setIssueSheetOpen(true);
-      setIssueDetailLoading(true);
-      setIssueTraceTasks(null);
-      setIssueActiveTask(null);
-      try {
-        const res = await getIssue(workspaceId, issueId);
-        setIssueDetail(res);
-        if (res.issue.latest_task_id) {
-          getTask(res.issue.latest_task_id, workspaceId)
-            .then((task) => setIssueActiveTask(task))
-            .catch(() => setIssueActiveTask(null));
-        }
-        if (res.issue.trace_id) {
-          getTrace(res.issue.trace_id, workspaceId)
-            .then((t) => setIssueTraceTasks(t.tasks))
-            .catch(() => setIssueTraceTasks(null));
-        }
-      } catch (err) {
-        toast.error(
-          err instanceof Error ? err.message : "Failed to load issue",
-        );
-        setIssueSheetOpen(false);
-      } finally {
-        setIssueDetailLoading(false);
-      }
-    },
-    [workspaceId],
-  );
-
-  const issueConvId = issueDetail?.issue?.conversation_id ?? null;
-  const issueTaskId = issueDetail?.issue?.latest_task_id ?? null;
 
   useEffect(() => {
     if (!issueSheetOpen || !selectedIssueId) return;
