+++ src/web/src/hooks/use-file-attachments.ts
@@ -1,0 +1,101 @@
+import React, { useCallback, useRef, useState } from "react";
+import { toast } from "sonner";
+
+const MAX_FILE_SIZE = 10 * 1024 * 1024; // 10 MB
+
+/**
+ * Owns the composer's file-staging UI: the pending (not-yet-sent) files, the
+ * hidden file-input ref, add/remove handlers, and drag-and-drop state.
+ *
+ * Extracted verbatim from agent-chat-view.tsx. The optimistic-send bookkeeping
+ * (`failedSends`, `pendingFilesByMessage`) intentionally stays OWNED by the
+ * component — send/retry read and write it — so this hook only manages the
+ * staging concern. `pendingFiles` + `setPendingFiles` are returned so the
+ * component's send/retry path (Tier 3) can clear/restore them exactly as before.
+ *
+ * Has no effects, so it carries no effect-order constraints.
+ */
+export function useFileAttachments() {
+  const [pendingFiles, setPendingFiles] = useState<File[]>([]);
+  const fileInputRef = useRef<HTMLInputElement>(null);
+
+  const addPendingFiles = useCallback((files: File[]) => {
+    const valid: File[] = [];
+    for (const file of files) {
+      if (file.size > MAX_FILE_SIZE) {
+        toast.error(`"${file.name}" exceeds 10 MB limit`);
+      } else {
+        valid.push(file);
+      }
+    }
+    if (valid.length > 0) {
+      setPendingFiles((prev) => [...prev, ...valid]);
+    }
+  }, []);
+
+  const handleFileSelect = useCallback(
+    (e: React.ChangeEvent<HTMLInputElement>) => {
+      const fileList = e.target.files;
+      if (!fileList) return;
+      addPendingFiles(Array.from(fileList));
+      // Reset input so re-selecting the same file works
+      e.target.value = "";
+    },
+    [addPendingFiles],
+  );
+
+  const removePendingFile = useCallback((index: number) => {
+    setPendingFiles((prev) => prev.filter((_, i) => i !== index));
+  }, []);
+
+  const [dragging, setDragging] = useState(false);
+  const dragCounter = useRef(0);
+
+  const handleDragEnter = useCallback((e: React.DragEvent) => {
+    e.preventDefault();
+    e.stopPropagation();
+    dragCounter.current++;
+    if (e.dataTransfer.types.includes("Files")) {
+      setDragging(true);
+    }
+  }, []);
+
+  const handleDragLeave = useCallback((e: React.DragEvent) => {
+    e.preventDefault();
+    e.stopPropagation();
+    dragCounter.current--;
+    if (dragCounter.current === 0) {
+      setDragging(false);
+    }
+  }, []);
+
+  const handleDragOver = useCallback((e: React.DragEvent) => {
+    e.preventDefault();
+    e.stopPropagation();
+  }, []);
+
+  const handleDrop = useCallback(
+    (e: React.DragEvent) => {
+      e.preventDefault();
+      e.stopPropagation();
+      setDragging(false);
+      dragCounter.current = 0;
+      addPendingFiles(Array.from(e.dataTransfer.files));
+    },
+    [addPendingFiles],
+  );
+
+  return {
+    pendingFiles,
+    setPendingFiles,
+    fileInputRef,
+    addPendingFiles,
+    handleFileSelect,
+    removePendingFile,
+    dragging,
+    handleDragEnter,
+    handleDragLeave,
+    handleDragOver,
+    handleDrop,
+  };
+}
